refactor: isolate ld-suites to individual files
