+++ packages/credential-w3c/package.json
@@ -26,7 +26,6 @@
     "debug": "^4.1.1",
     "did-jwt-vc": "2.1.7",
     "did-resolver": "3.1.3",
-    "ecdsa-secp256k1-signature-2019": "^1.0.1",
     "jsonld": "^5.2.0",
     "jsonld-signatures": "^9.0.2",
     "uint8arrays": "^2.1.3",
